requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
attrs==25.4.0
black==23.3.0
brotli
certifi
cffi
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
cmake==4.1.2
colorama
contourpy==1.3.2
crc32c==2.7.1
cycler==0.12.1
Cython==3.2.1
d4rl==1.1
diffusers==0.33.1
-e git+https://github.com/real-stanford/diffusion_policy.git@5ba07ac6661db573af695b419a7947ecb704690f#egg=diffusion_policy
dill==0.4.0
distlib==0.3.9
dm-env==1.6
dm-tree==0.1.9
dm_control==1.0.34
docker-pycreds==0.4.0
donfig==0.8.1.post1
egl-probe>=1.0.2
einops==0.8.1
etils==1.13.0
Farama-Notifications==0.0.4
fasteners==0.20
filelock==3.18.0
flake8==6.1.0
fonttools==4.58.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
glfw==2.10.0
gmpy2
grpcio==1.76.0
gym==0.23.1
gym-notices==0.0.8
gym-pusht==0.1.0
gymnasium==0.29.1
gymnasium-robotics>=1.2.0,<1.3.0
h2
h5py==3.15.1
hpack
huggingface-hub==0.31.2
hydra-core==1.3.2
hyperframe
identify==2.6.10
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
isort==5.12.0
jinja2
kiwisolver==1.4.8
labmaze==1.0.6
lazy_loader==0.4
lightning
lightning-utilities
llvmlite==0.44.0
lxml==6.0.2
Markdown==3.10
markupsafe
matplotlib==3.10.3
mccabe==0.7.0
mpmath
mujoco==3.3.7
mujoco-py==2.1.2.14
mypy_extensions==1.1.0
networkx==3.4.2
nodeenv==1.9.1
numba==0.61.2
numcodecs<=0.13.1
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.11.0.86
optree
packaging
pandas==2.2.3
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
protobuf==6.31.0
psutil==7.0.0
pybind11
pybind11_global
pybullet==3.2.7
pycodestyle==2.11.1
pycparser
pydantic==2.11.4
pydantic_core==2.33.2
pyflakes==3.1.0
pygame==2.5.2
Pygments==2.19.1
pymunk==6.11.1
PyOpenGL==3.1.10
pyparsing==3.2.3
pysocks
pytest==8.4.0
python-dateutil==2.9.0.post0
pytorch-lightning
pytz==2025.2
pyyaml
regex==2024.11.6
requests
robomimic==0.3.0
safetensors==0.5.3
scikit-image==0.22.0
scikit-misc==0.3.0
scipy==1.15.3
sentry-sdk==2.28.0
setproctitle==1.3.6
shapely==2.1.2
six==1.17.0
smmap==5.0.2
sympy
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
termcolor==3.2.0
tifffile==2025.5.10
torch
torchaudio==2.1.2
torchmetrics
torchtyping==0.1.5
torchvision==0.16.2
tqdm
typeguard==2.13.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3
virtualenv==20.31.2
wandb==0.19.11
Werkzeug==3.1.3
wrapt==2.0.1
zarr<3.5
zipp==3.21.0
zstandard==0.25.0
av
pettingzoo
# widowx
pyquaternion
funcsigs
moviepy<2
decorator
git+https://github.com/youliangtan/edgeml.git